requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
bitarray==2.9.2
bleak==0.22.2
certifi==2024.7.4
charset-normalizer==3.3.2
ckzg==2.0.0
confluent-kafka==2.5.3
crypto_cpp_py==1.4.4
cytoolz==0.12.3
datadog==0.50.0
dateparser==1.2.0
dbus-fast==2.22.1
ecdsa==0.18.0
ECPy==1.2.5
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==2.3.1
eth_abi==5.1.0
exceptiongroup==1.2.2
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
hexbytes==0.3.1
hidapi==0.14.0.post2
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lark==1.2.2
ledgerblue==0.1.48
ledgereth==0.9.1
libusb1==3.1.0
lru-dict==1.2.0
marshmallow==3.22.0
marshmallow-oneofschema==3.0.1
marshmallow_dataclass==8.6.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
ndeflib==0.3.3
nfcpy==1.0.4
numpy==2.1.0
packaging==24.1
paradex_py==0.4.2
parsimonious==0.10.0
pillow==10.4.0
poseidon_py==0.1.4
protobuf==5.27.3
pycryptodome==3.20.0
pycryptodomex==3.20.0
pyDes==2.0.1
pyelftools==0.31
pyserial==3.5
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-gnupg==0.5.2
python-u2flib-host==3.0.3
pytz==2024.1
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rlp==3.0.0
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starknet-crypto-py==0.2.0
starknet-py==0.22.0
sympy==1.11.1
toolz==0.12.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
web3==6.20.2
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
